Create utils folder, split up utils.py
